Fix off by one error with EndOfBackendKeys
It looks like the alias was cargo culted from functionality keys,
but you only need to actually do an alias if you need a break in the
middle of the sequence while keeping the numbering contiguous, which
doesn't apply here.  (It would also be necessary if we exactly filled
the representation type, but uint8_t has plenty of space.)

I audited all uses of this key and it turns out that it had committed
(harmless) off by one errors.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: 2a83615a3ed245dd117f28b78deb088007027428
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81712