[vulkan] Only use packed context for State of VulkanOpContext
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82144


Previously, `VulkanOpContext` uses a tuple of `c10::GenericList` objects as its `State`, which contains the "unpacked" (original) and "packed" (modified) arguments for an op. However, since the packed arguments contain converted `vTensor` objects, this causes issues when serializing Vulkan models which contain `VulkanOpContext` objects.

This diff redefines the `State` of the `VulkanOpContext` class to be a single `c10::GenericList` that corresponds to the "unpacked" arguments (now renamed to the "source" arguments) in order to faciliate model serialization.
ghstack-source-id: 162487625

Differential Revision: [D38124961](https://our.internmc.facebook.com/intern/diff/D38124961/)