Generate vmap plumbing on all native_functions, not just ones in allowlist
Motivation
- The initial motivation for the allowlist is that we were checking in
VmapGeneratedPlumbing.h to pytorch/functorch but people were changing
schemas of operators in pytorch/pytorch. The allowlist helped reduce the
number of collisions (because people change schemas of more operators
than we had in the allowlist). This is no longer a problem because
functorch is in the pytorch/pytorch repo
- Avoid merge conflicts. Multiple people editing the allowlist leads to
merge conflicts; getting rid of that alleviates it.

Test Plan:
- wait for CI

ghstack-source-id: 1c1b2696f670ce211bcd322503d706d281f9c5f3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82319